+++ docs/demo/react/index.js
@@ -65,6 +65,7 @@ const videoJsOptions = {
             audio: true,
             video: true,
             maxLength: 10,
+            displayMilliseconds: false,
             debug: true
         }
     }