+++ test/plugins/webm-wasm-plugin.spec.js
@@ -2,14 +2,14 @@
  * @since 3.5.0
  */
 
-import TestHelpers from '../test-helpers.js';
+import TestHelpers from '../test-helpers';
 
-import Event from '../../src/js/event.js';
+import Event from '../../src/js/event';
 
 // registers the plugin
-import {RecordRTCEngine} from '../../src/js/engine/record-rtc.js';
-import WebmWasmEngine from '../../src/js/plugins/webm-wasm-plugin.js';
-import {WEBMWASM} from '../../src/js/engine/record-engine.js';
+import {RecordRTCEngine} from '../../src/js/engine/record-rtc';
+import WebmWasmEngine from '../../src/js/plugins/webm-wasm-plugin';
+import {WEBMWASM} from '../../src/js/engine/record-engine';
 
 
 /** @test {WebmWasmEngine} */
@@ -28,6 +28,7 @@ describe('plugins.webm-wasm-plugin', () => {
         player.one(Event.FINISH_RECORD, () => {
             // received a blob file
             expect(player.recordedData instanceof Blob).toBeTruthy();
+            expect(player.recordedData.name).toEndWith('.webm');
 
             // wait till it's loaded before destroying
             // (XXX: create new event for this)